% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drinner.R
\name{drinner}
\alias{drinner}
\title{Inner loop of Seasonal Trend Decomposition of Time Series by Loess}
\usage{
drinner(Inner_input, Inner_output, n, n.p, vari, cyctime, seaname, s.window,
  s.degree, t.window, t.degree, l.window, l.degree, s.jump, t.jump, l.jump,
  critfreq, s.blend, t.blend, l.blend, crtI, sub.labels, sub.start, infill,
  Clcontrol)
}
\arguments{
\item{Inner_input}{the HDFS path of input files for inner loop}

\item{Inner_output}{the HDFS path of output files for inner loop}

\item{n}{the number of all observation in the series}

\item{n.p}{periodicity of the seasonal component. In R's \code{stl} function, this is the frequency of the time series.}

\item{s.window}{either the character string \code{"periodic"} or the span (in lags) of the loess window for seasonal extraction, which should be odd. This has no default.}

\item{s.degree}{degree of locally-fitted polynomial in seasonal extraction. Should be 0, 1, or 2.}

\item{t.window}{the span (in lags) of the loess window for trend extraction, which should be odd. If \code{NULL}, the default, \code{nextodd(ceiling((1.5*period) / (1-(1.5/s.window))))}, is taken.}

\item{t.degree}{degree of locally-fitted polynomial in trend extraction. Should be 0, 1, or 2.}

\item{l.window}{the span (in lags) of the loess window of the low-pass filter used for each subseries. Defaults to the smallest odd integer greater than or equal to \code{n.p} which is recommended since it prevents competition between the trend and seasonal components. If not an odd integer its given value is increased to the next odd one.}

\item{l.degree}{degree of locally-fitted polynomial for the subseries low-pass filter. Should be 0, 1, or 2.}

\item{s.jump, t.jump, l.jump}{integers at least one to increase speed of the respective smoother. Linear interpolation happens between every \code{*.jump}th value.}

\item{critfreq}{the critical frequency to use for automatic calculation of smoothing windows for the trend and high-pass filter.}

\item{s.blend, t.blend, l.blend}{vectors of proportion of blending to degree 0 polynomials at the endpoints of the series.}

\item{crtI}{index of corrent inner loop}

\item{sub.labels}{optional vector of length n.p that contains the labels of the subseries in their natural order (such as month name, day of week, etc.), used for strip labels when plotting. All entries must be unique.}

\item{sub.start}{which element of sub.labels does the series begin with. See details.}

\item{Clcontrol}{a list contains all tuning parameters for the mapreduce job.}
}
\description{
The inner loop of STL. A new implementation of STL. Allows for NA values, local quadratic smoothing, post-trend smoothing, and endpoint blending. The usage is very similar to that of R's built-in \code{stl()}.
Allows seasonal fitting in parallel.
}
\details{
The seasonal component is found by \emph{loess} smoothing the seasonal sub-series (the series of all January values, \ldots); if \code{s.window = "periodic"} smoothing is effectively replaced by taking the mean. The seasonal values are removed, and the remainder smoothed to find the trend. The overall level is removed from the seasonal component and added to the trend component. This process is iterated a few times. The \code{remainder} component is the residuals from the seasonal plus trend fit.
Cycle-subseries labels are useful for plotting and can be specified through the sub.labels argument. Here is an example for how the sub.labels and sub.start parameters might be set for one situation. Suppose we have a daily series with n.p=7 (fitting a day-of-week component). Here, sub.labels could be set to c("Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"). Now, if the series starts with a Wednesday value, then one would specify sub.labels=4, since Wednesday is the fourth element of sub.labels. This ensures that the labels in the plots to start the plotting with Sunday cycle-subseries instead of Wednesday.
}
\note{
This is a complete re-implementation of the STL algorithm, with the loess part in C and the rest in R. Moving a lot of the code to R makes it easier to experiment with the method at a very minimal speed cost. Recoding in C instead of using R's built-in loess results in better performance, especially for larger series.
}
\author{
Xiaosu Tong, Ryan Hafen
}
\references{
R. B. Cleveland, W. S. Cleveland, J. E. McRae, and I. Terpenning (1990) STL: A Seasonal-Trend Decomposition Procedure Based on Loess. \emph{Journal of Official Statistics}, \bold{6}, 3--73.
}

