% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drrobust.R
\name{drrobust}
\alias{drrobust}
\title{Calculate the robust weight for STL}
\usage{
drrobust(input, output, vari, zero.weight)
}
\arguments{
\item{input}{The path of input sequence file on HDFS. It should be by location division.}

\item{output}{The path of output sequence file on HDFS. It is by time division.}

\item{zero.weight}{all parameters that are needed for space-time fitting}

\item{seasonal}{The number of the reduce tasks.}

\item{trend}{Logical argument, if TRUE, then the elevation attribute from input value data.frame is log2 transformation
If FALSE, a log2 transformation is added to the elevation attribute.}
}
\description{
Calculate the robust weight for each observation of the time series
}
\details{
swaptoTime is used for switch division by location to division by time.
}
\examples{
    FileInput <- "/ln/tongx/Spatial/tmp/tmax/a1950/byseason.season"
    FileOutput <- "/ln/tongx/Spatial/tmp/tmax/a1950/byseason.outer"
    \dontrun{
       drrobust(FileInput, FileOutput, vari="resp", zero.weight=)
    }
}
\author{
Xiaosu Tong
}

