% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapreduce.control.R
\name{mapreduce.control}
\alias{mapreduce.control}
\title{Set mapreduce parameter for drSpaceTime fitting}
\usage{
mapreduce.control(reduceTask = 0, spill.percent = 0.8, io.sort = 256,
  libLoc = NULL)
}
\arguments{
\item{reduceTask}{The reduce task number, also the number of output files. If set to be 0, then there is no shuffle and sort stage after map.}

\item{spill.percent}{The threshold usage proportion for both the map output memory buffer and record boundaries index to start the process of spilling to disk.}

\item{io.sort}{The size, in megabytes, of the memory buffer to use while sorting map output.}
}
\value{
A list with space-time fitting parameters.
}
\description{
Set control parameters of mapreduce for drSpaceTime fits
}
\examples{
    mapreduce.control()
}
\author{
Xiaosu Tong
}

