% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predNew_local.R
\name{predNew_local}
\alias{predNew_local}
\title{Prediction at new locations based on the fitting results in memory.}
\usage{
predNew_local(original, newdata, mlcontrol = spacetime.control())
}
\arguments{
\item{original}{The data.frame which contains all fitting results of original dataset. The data.frame
is saved in memory, not on HDFS.}

\item{newdata}{A data.frame includes all locations' longitude, latitude, and elevation,
where the prediction is to be calculated.}

\item{mlcontrol}{Should be a list object generated from \code{spacetime.control} function.
The list including all necessary smoothing parameters of nonparametric fitting.}
}
\description{
The prediction at new locations are calculated based on the fitting results
saved in memory based on the original dataset.
}
\examples{
    library(maps)
    library(Spaloess)

    new.grid <- expand.grid(
      lon = seq(-126, -67, by = 1),
      lat = seq(25, 49, by = 1)
    )
    instate <- !is.na(map.where("state", new.grid$lon, new.grid$lat))
    new.grid <- new.grid[instate, ]

    elev.fit <- spaloess( elev ~ lon + lat,
      data = station_info,
      degree = 2,
      span = 0.05,
      distance = "Latlong",
      normalize = FALSE,
      napred = FALSE,
      alltree = FALSE,
      family="symmetric",
      control=loess.control(surface = "direct")
    )
    grid.fit <- predloess(
      object = elev.fit,
      newdata = data.frame(
        lon = new.grid$lon,
        lat = new.grid$lat
      )
    )
    new.grid$elev <- grid.fit

    n <- 1000 # just use 1000 stations as example
    set.seed(99)
    first_stations <- sample(unique(tmax_all$station.id), n)
    small_dt <- subset(tmax_all, station.id \%in\% first_stations)
    small_dt$station.id <- as.character(small_dt$station.id)
    small_dt$month <- as.character(small_dt$month)
    mlcontrol <- spacetime.control(
      vari="tmax", time="date", n=576, n.p=12, stat_n=n, surf = "interpolate",
      s.window="periodic", t.window = 241, degree=2, span=0.15, Edeg=0
    )

    fitted <- drsstl(
      data=small_dt,
      output=NULL,
      model_control=mlcontrol
    )
    rst <- predNew_local(
      original = fitted, newdata = new.grid, mlcontrol = mlcontrol
    )
}
\author{
Xiaosu Tong
}
\seealso{
\code{\link{spacetime.control}}, \code{\link{mapreduce.control}}
}

