% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repbyTime.R
\name{repbyTime}
\alias{repbyTime}
\title{Swap to division by time}
\usage{
repbyTime(input, output, Rep = 8000, buffSize = 7, reduceTask,
  control = spacetime.control())
}
\arguments{
\item{input}{The path of input sequence file on HDFS. It should be by location division.}

\item{output}{The path of output sequence file on HDFS. It is by location division but duplicate the time series to be longer.}

\item{Rep}{The replication time for the time series of each location}

\item{buffSize}{To control how many location will be send to one mapper}

\item{reduceTask}{The number of the reduce tasks.}
}
\description{
Switch input key-value pairs which is division by location
to the key-value pairs which is division by time.
}
\details{
swaptoTime is used for switch division by location to division by time.
}
\author{
Xiaosu Tong
}

