% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustWeight.R
\name{robustWeight}
\alias{robustWeight}
\title{Calculate the robust weight for STL}
\usage{
robustWeight(n, Y, seasonal, trend, zero.weight)
}
\arguments{
\item{n}{The path of input sequence file on HDFS. It should be by location division.}

\item{Y}{The path of output sequence file on HDFS. It is by time division.}

\item{seasonal}{The number of the reduce tasks.}

\item{trend}{Logical argument, if TRUE, then the elevation attribute from input value data.frame is log2 transformation
If FALSE, a log2 transformation is added to the elevation attribute.}

\item{zero.weight}{all parameters that are needed for space-time fitting}
}
\description{
Calculate the robust weight for each observation of the time series
}
\details{
swaptoTime is used for switch division by location to division by time.
}
\author{
Xiaosu Tong
}

