% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetime.control.R
\name{spacetime.control}
\alias{spacetime.control}
\title{Set smoothing parameters for drSpaceTime fitting}
\usage{
spacetime.control(vari = "resp", time = "date", seaname = "month", n,
  n.p = 12, s.window, s.degree = 1, t.window = NULL, t.degree = 1,
  inner = 4, outer = 1, statbytime = 2, degree, span, Edeg,
  surf = c("interpolate", "direct"), iterations = 1)
}
\arguments{
\item{vari}{variable name in string of the response variable}

\item{time}{variable name in string of time index of the whole time series}

\item{seaname}{variable name in string of the seasonal variable}

\item{n}{the number of total observations}

\item{n.p}{the number of observation in each subseries}

\item{s.window}{either the character string \code{"periodic"} or the span (in lags) of 
the loess window for seasonal extraction, which should be odd. This has no default.}

\item{s.degree}{degree of locally-fitted polynomial in seasonal extraction. Should be 0, 1, or 2.}

\item{inner}{The iteration time for inner loop of stlplus for time dimension fitting}

\item{outer}{The iteration time for outer loop of stlplus for time dimension fitting}

\item{statbytime}{The number of locations will be grouped together in the by time division after stlfit.
The parameter is only used for \code{swaptoTime}. Since there may be to many time
point in each location, the \code{swaptoTime} looping all time point will add to much
overhead caused by \code{rhcollect}. So every \code{statbytime} time point are collect
into one key-value pair.}

\item{degree}{smoothing degree for the spatial loess smoothing. It can be 0, 1, or 2.}

\item{span}{smoothing span for the spatial loess smoothing.}

\item{Edeg}{the degree for the conditioanl parametric model including elevation.}

\item{surf}{should the fitted surface be computed exactly or via interpolation from a kd tree.}

\item{iterations}{the number of iterations used for the space-time back-fitting.}
}
\value{
A list with space-time fitting parameters.
}
\description{
Set control parameters for the smoothing fit
}
\examples{
    spacetime.control(
     n = 786432, n.p = 12, s.window = 21, s.degree = 1, t.window = 241, 
      t.degree = 1, degree = 2, span = 0.015, Edeg = 2, surf = "interpolate"
    )
}
\author{
Xiaosu Tong
}
\references{
R. B. Cleveland, W. S. Cleveland, J. E. McRae, and I. Terpenning (1990) STL: A Seasonal-Trend Decomposition Procedure Based on Loess. \emph{Journal of Official Statistics}, \bold{6}, 3--73.
}

