% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaofit.R
\name{spaofit}
\alias{spaofit}
\title{Apply the spatial loess fitting to the original observations at 
all locations in each month in parallel}
\usage{
spaofit(input, output, info, model_control = spacetime.control(),
  cluster_control = mapreduce.control())
}
\arguments{
\item{input}{The path of input sequence file on HDFS. It should be by-month division.}

\item{output}{The path of output sequence file on HDFS. It is also by-month division but with 
seasonal and trend components}

\item{info}{The RData on HDFS which contains all station metadata. Make sure
copy the RData of station_info to HDFS first using rhput.}

\item{model_control}{Should be a list object generated from \code{spacetime.control} function.
The list including all necessary smoothing parameters of nonparametric fitting.}

\item{cluster_control}{Should be a list object generated from \code{mapreduce.control} function.
The list including all necessary Rhipe parameters and also user tunable 
MapReduce parameters.}
}
\description{
Call \code{spaloess} function on the spatial domain in each month in parallel.
Every spatial domain uses the same smoothing parameters. NA observations will be
imputed.
}
\examples{
    FileInput <- "/tmp/bymth"
    FileOutput <- "/tmp/bymthfit"
    ccontrol <- mapreduce.control(libLoc=NULL, reduceTask=0)
    mcontrol <- spacetime.control(
      vari="resp", time="date", n=576, n.p=12, stat_n=7738,
      s.window=13, t.window = 241, degree=2, span=0.015, Edeg=2
    )

    spaofit(
      FileInput, FileOutput,
      info="/tmp/station_info.RData", 
      model_control=mcontrol, cluster_control=ccontrol
    )
}
\author{
Xiaosu Tong
}

