% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstl_local.R
\name{sstl_local}
\alias{sstl_local}
\title{Apply sstl routine to a data.frame of spatial-temporal dataset in the memory.}
\usage{
sstl_local(data, mlcontrol = spacetime.control())
}
\arguments{
\item{data}{The input data.frame which contains observation, lon, lat, elev, and year, month}

\item{mlcontrol}{Should be a list object generated from \code{spacetime.control} function.
The list including all necessary smoothing parameters of nonparametric fitting.}
}
\description{
Assuming data has been read into the memory as a data.frame. Each row of the data.frame
contains the observation in a given month at given location. Month index and station
location information are saved in 5 columns of the data.frame. The name of these columns
should be "lon", "lat", "elev", "year", and "month".
}
\examples{
    head(tmax_all)
    mcontrol <- spacetime.control(
      vari="tmax", n=576, n.p=12, stat_n=7738,
      s.window=13, t.window = 241, degree=2, span=0.015, Edeg=2
    )
    sstl_local(tmax_all, mlcontrol=mcontrol)
}
\author{
Xiaosu Tong
}

