% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstl_mr.R
\name{sstl_mr}
\alias{sstl_mr}
\title{Apply sstl routine to dataset saved on HDFS}
\usage{
sstl_mr(input, output, stat_info, mlcontrol = spacetime.control(),
  clcontrol = mapreduce.control())
}
\arguments{
\item{input}{The input path of raw text file on HDFS}

\item{mlcontrol}{Should be a list object generated from \code{spacetime.control} function.
The list including all necessary smoothing parameters of nonparametric fitting.}

\item{clcontrol}{Should be a list object generated from \code{mapreduce.control} function.
The list including all necessary Rhipe parameters and also user tunable
MapReduce parameters.}
}
\description{
Input raw text file on HDFS is the original input dataset. After a series of MapReudce
jobs, the final fitting results are saved as output_bymth and output_bystat subdirectory
inside of output path on HDFS.
}
\examples{
\dontrun{
    mcontrol <- spacetime.control(
      vari="tmax", n=576, n.p=12, stat_n=7738,
      s.window=13, t.window = 241, degree=2, span=0.015, Edeg=2
    )
    ccontrol <- mapreduce.control(
      libLoc= NULL, reduceTask=169, io_sort=512, BLK=128, slow_starts = 0.5,
      map_jvm = "-Xmx200m", reduce_jvm = "-Xmx200m",
      map_memory = 1024, reduce_memory = 1024,
      reduce_input_buffer_percent=0.4, reduce_parallelcopies=10,
      reduce_merge_inmem=0, task_io_sort_factor=100,
      spill_percent=0.9, reduce_shuffle_input_buffer_percent = 0.8,
      reduce_shuffle_merge_percent = 0.4
    )
    sstl_mr(
      input = "/tmp/tmax.txt", output = "/tmp/output",
      stat_info = "/tmp/station_info.RData", mlcontrol = mcontrol,
      clcontrol = clcontrol
    )
}
}
\author{
Xiaosu Tong
}

