% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drsstl-package.R
\docType{data}
\name{station_info}
\alias{station_info}
\title{Metadata of Stations}
\source{
All station metadata: \url{http://www.image.ucar.edu/Data/US.monthly.met/FullData.shtml#temp}
}
\description{
Metadata for all stations including longitude, latitude, elevation, station.id
}
\keyword{data}

