% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stlfit.R
\name{stlfit}
\alias{stlfit}
\title{Conduct the stlplus fitting at each location in parallel}
\usage{
stlfit(input, output, model_control = spacetime.control(),
  cluster_control = mapreduce.control())
}
\arguments{
\item{input}{The path of input sequence file on HDFS. It should be by location division.}

\item{output}{The path of output sequence file on HDFS. It is by location division but with seasonal and trend components}

\item{model_control}{The list contains all smoothing parameters}

\item{cluster_control}{A list contains all mapreduce tuning parameters.}
}
\description{
call stlplus function on time series at each location in parallel.
Every station uses the same smoothing parameter
}
\examples{
    FileInput <- "/wsc/tongx/Spatial/tmp/tmax/simulate/bystation"
    FileOutput <- "/wsc/tongx/Spatial/tmp/tmax/simulate/bystation.stlfit"
    me <- spacetime.control(libLoc=lib.loc)
    you <- mapreduce.control(vari="resp", time="date", seaname="month", n=3340800, n.p=12, s.window=13, t.window = 241)
    \dontrun{
      stlfit(FileInput, FileOutput, me, you)
    }
}
\author{
Xiaosu Tong
}

