% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swaptoLoc.R
\name{swaptoLoc}
\alias{swaptoLoc}
\title{Swap to division by location}
\usage{
swaptoLoc(input, output, sub, cluster_control = mapreduce.control())
}
\arguments{
\item{input}{The path of input sequence file on HDFS. It should be by location division.}

\item{output}{The path of output sequence file on HDFS. It is by time division.}

\item{sub}{The number of different locations within one hadoop key-value pair.}

\item{cluster_control}{A list contains all mapreduce tuning parameters.}

\item{model_control}{The list contains all smoothing parameters}
}
\description{
Switch input key-value pairs which is division by time
to the key-value pairs which is division by location.
}
\examples{
    FileInput <- "/wsc/tongx/spatem/tmax/sim/bymthfit256"
    FileOutput <- "/wsc/tongx/spatem/tmax/sim/bystat256"
    me <- mapreduce.control(libLoc="/home/tongx/R_LIBS", io_sort=512, BLK=256, reduce_input_buffer_percent=0.7, reduce_merge_inmem=0, task_io_sort_factor=100, spill_percent=1)
    \dontrun{
      swaptoLoc(FileInput, FileOutput, sub=1, cluster_control=me)
    }
}
\author{
Xiaosu Tong
}

