% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swaptoSeason.R
\name{swaptoSeason}
\alias{swaptoSeason}
\title{Swap to division by season}
\usage{
swaptoSeason(input, output, reduceTask = 0, control = spacetime.control())
}
\arguments{
\item{input}{The path of input sequence file on HDFS. It should be by location.}

\item{output}{The path of output sequence file on HDFS. It is by season division.}

\item{reduceTask}{The number of the reduce tasks.}

\item{control}{all parameters that are needed for space-time fitting}
}
\description{
Switch input key-value pairs which is division by location
to the key-value pairs which is division by season.
}
\details{
swaptoSeason is used for switch division by location to division by season.
    The output files will be intput to parallel seasonal fitting of stlplus.
}
\examples{
    FileInput <- "/ln/tongx/Spatial/a1950/bystation"
    FileOutput <- "/ln/tongx/Spatial/a1950/byseason"
    \dontrun{
      swaptoSeason(FileInput, FileOutput, control=spacetime.control(libLoc=.libPaths()))
    }
}
\author{
Xiaosu Tong
}

