% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swaptoSubser.R
\name{swaptoSubser}
\alias{swaptoSubser}
\title{Swap to division by subseries}
\usage{
swaptoSubser(input, output, cluster_control, model_control)
}
\arguments{
\item{input}{The path of input file on HDFS. It should be by location.}

\item{output}{The path of output file on HDFS. It is by subseries division.}

\item{cluster_control}{Should be a list object generated from \code{mapreduce.control} function.
The list including all necessary Rhipe parameters and also user tunable 
MapReduce parameters.}

\item{model_control}{Should be a list object generated from \code{spacetime.control} function.
The list including all necessary smoothing parameters of nonparametric fitting.}
}
\description{
Switch input key-value pairs which is division by location
to the key-value pairs which is division by subseries.
}
\details{
\code{swaptoSubser} is used for switch division by location to division by subseries.
    The smoothing procedure can be applied to each subseries in parallel. So the
     output files will be intput to parallel seasonal fitting of stlplus.
}
\examples{
    FileInput <- "/wsc/tongx/spatem/tmax/sims/bystat"
    FileOutput <- "/wsc/tongx/spatem/tmax/sims/bysubser"
    ccontrol <- mapreduce.control(libLoc=.libPaths(), reduceTask=179)
    mcontrol <- spacetime.control(
      vari = "resp", time = "date", seaname = "month", 
      n = 786432, n.p = 12, s.window = "periodic", t.window = 241, 
      degree = 2, span = 0.015, Edeg = 2, statbytime = 2
    )
    swaptoSubser(FileInput, FileOutput, cluster_control=ccontrol, model_control=mcontrol)
}
\author{
Xiaosu Tong
}
\seealso{
\code{\link{spacetime.control}}, \code{\link{mapreduce.control}}
}

