% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swaptoTime.R
\name{swaptoTime}
\alias{swaptoTime}
\title{Spatially Local Polynomial Regression Fitting}
\usage{
swaptoTime(input, output, elevFlag = TRUE, reduceTask)
}
\arguments{
\item{input}{The path of input sequence file on HDFS. It should be by location division.}

\item{output}{The path of output sequence file on HDFS. It is by time division.}

\item{elevFlag}{Logical argument, if TRUE, then the elevation attribute from input value data.frame is log2 transformation
If FALSE, a log2 transformation is added to the elevation attribute.}

\item{reduceTask}{The number of the reduce tasks.}
}
\description{
The first layer of the Spatial locally weighted regression, using local fitting with different 
type of distance calculation.
}
\details{
swaptoTime is used for switch division by location to division by time.
}
\examples{
    FileInput <- "/ln/tongx/Spatial/a1950/bystation"
    FileOutput <- "/ln/tongx/Spatial/a1950/bymonth"
    \dontrun{
      swaptoTime(FileInput, FileOutput, elevFlag=TRUE)
    }
}
\author{
Xiaosu Tong
}

