% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drSpaceTime-package.R
\docType{data}
\name{tmax_all}
\alias{tmax_all}
\title{DATA TITLE}
\source{
0-5 years: \url{http://www.who.int/childgrowth/software/en/}

5-15 years: \url{http://www.who.int/growthref/tools/en/}
}
\description{
DESCRIPTION
}
\keyword{data}

