% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drrobust.R
\name{drrobust}
\alias{drrobust}
\title{Calculate the robust weight for STL}
\usage{
drrobust(input, output, vari, Clcontrol, zero.weight = 1e-06)
}
\arguments{
\item{input}{The path of input sequence file on HDFS. It should be by location division.}

\item{output}{The path of output sequence file on HDFS. It is by time division.}

\item{vari}{variable name in string of the response variable}

\item{zero.weight}{value to use as zero for zero weighting}

\item{reduceTask}{The reduce task number, also the number of output files. If set to be 0, then there is no shuffle and sort stage after map.}

\item{spill.percent}{The threshold usage proportion for both the map output memory buffer and record boundaries index to start the process of spilling to disk.}

\item{io.sort}{The size, in megabytes, of the memory buffer to use while sorting map output.}
}
\description{
Calculate the robust weight for each observation of the time series
}
\details{
Calculate the robust weight for each location
}
\examples{
    FileInput <- "/ln/tongx/Spatial/tmp/tmax/a1950/byseason.inner"
    FileOutput <- "/ln/tongx/Spatial/tmp/tmax/a1950/byseason.outer"
    \dontrun{
       drrobust(FileInput, FileOutput, vari="resp")
    }
}
\author{
Xiaosu Tong
}

