% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIn.R
\name{readIn}
\alias{readIn}
\title{Readin Raw text data files and save it as by time division on HDFS.}
\usage{
readIn(input, output, info, cluster_control = mapreduce.control(),
  cshift = 1)
}
\arguments{
\item{input}{The path of input file on HDFS. It should be raw text file.}

\item{output}{The path of output file on HDFS. It is by time division.}

\item{info}{The RData on HDFS which contains all station metadata. Make sure
copy the RData of station_info to HDFS first using rhput.}

\item{cluster_control}{all parameters that are needed for mapreduce job}
}
\description{
Input raw text data file is divided into by time subsets and saved on HDFS
}
\examples{
    FileInput <- "/tmp/tmax.txt"
    FileOutput <- "/tmp/bymth"
    ccontrol <- mapreduce.control(
      libLoc=lib.loc, reduceTask=95, io_sort=100, slow_starts = 0.5,
      map_jvm = "-Xmx200m", reduce_jvm = "-Xmx200m", 
      map_memory = 1024, reduce_memory = 1024,
      reduce_input_buffer_percent=0.9, reduce_parallelcopies=5,
      spill_percent=0.9, reduce_shuffle_input_buffer_percent = 0.9,
      reduce_shuffle_merge_percent = 0.5
    )
    \dontrun{
      rhput("./station_info.RData", "/tmp/station_info.RData")
    }
    readIn(
      FileInput, FileOutput, info="/tmp/station_info.RData", cluster_control=ccontrol
    )
}
\author{
Xiaosu Tong
}

