% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIn.R
\name{readIn}
\alias{readIn}
\title{Readin Raw text data files and save it as by time division on HDFS.}
\usage{
readIn(input, output, info, cluster_control = mapreduce.control())
}
\arguments{
\item{input}{The path of input sequence file on HDFS. It should be raw text file.}

\item{output}{The path of output sequence file on HDFS. It is by time division.}

\item{cluster_control}{all parameters that are needed for mapreduce job}
}
\description{
Input raw text data file is divided into by time subsets and saved on HDFS
}
\examples{
   me <- mapreduce.control(
     libLoc="/home/tongx/R_LIBS", reduceTask=1300, io_sort=512, BLK=256, 
     reduce_input_buffer_percent=0.99, reduce_parallelcopies=10, 
     reduce_merge_inmem=0, task_io_sort_factor=100, 
     spill_percent=0.99, reduce_shuffle_input_buffer_percent = 0.9,
     reduce_shuffle_merge_percent = 0.99
   )
   \dontrun{
     readIn("/wsc/tongx/spatem/nRaw/tmax","/wsc/tongx/spatem/tmax/sim/bymth", info="/wsc/tongx/spatem/stationinfo/a1950UStinfo.RData", me) 
   }
}
\author{
Xiaosu Tong
}

