% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repbyTime.R
\name{repbyTime}
\alias{repbyTime}
\title{Swap to division by time}
\usage{
repbyTime(input, output, Rep = 5800, control = mapreduce.control())
}
\arguments{
\item{input}{The path of input sequence file on HDFS. It should be by location division.}

\item{output}{The path of output sequence file on HDFS. It is by location division but duplicate the time series to be longer.}

\item{Rep}{The replication time for the time series of each location}

\item{control}{A list contains all mapreduce tuning parameters.}
}
\description{
Switch input key-value pairs which is division by location
to the key-value pairs which is division by time.
}
\details{
repbyoTime is used for duplicate the time series at each location.
}
\author{
Xiaosu Tong
}

