% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repbyTime.R
\name{repbyTime}
\alias{repbyTime}
\title{Replicate the data in time dimension.}
\usage{
repbyTime(input, output, Rep = 5800, control = mapreduce.control())
}
\arguments{
\item{input}{The path of input sequence file on HDFS. It should be by location division.}

\item{output}{The path of output sequence file on HDFS. It is by location division but duplicate the time series to be longer.}

\item{Rep}{The replication time for the time series of each location}

\item{control}{A list contains all mapreduce tuning parameters.}
}
\description{
For simulation purpose to enlarge the data for each location.
}
\details{
repbyoTime is used for duplicate the time series at each location.
}
\examples{
    FileInput <- "/wsc/tongx/Spatial/tmp/tmax/simulate/bystation.orig"
    FileOutput <- "/wsc/tongx/Spatial/tmp/tmax/simulate/bystation.small"
    \dontrun{
      repbyTime(FileInput, FileOutput, Rep=5800, me)
    }
}
\author{
Xiaosu Tong
}

