% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repbyTime.R
\name{repbyTime}
\alias{repbyTime}
\title{Replicate the data in time dimension.}
\usage{
repbyTime(input, output, Srep, control = mapreduce.control())
}
\arguments{
\item{input}{The path of input sequence file on HDFS. It should be by location division.}

\item{output}{The path of output sequence file on HDFS. It is by location division but duplicate the time series to be longer.}

\item{control}{A list contains all mapreduce tuning parameters.}

\item{Rep}{The replication time for the time series of each location}
}
\description{
For simulation purpose to enlarge the data for each location.
}
\details{
repbyoTime is used for duplicate the time series at each location.
}
\examples{
    FileInput <- "/wsc/tongx/spatem/tmax/sim/bystation.orig"
    FileOutput <- "/wsc/tongx/spatem/tmax/sim/bystation"
    me <- mapreduce.control(libLoc=lib.loc, BLK = 256)
    \dontrun{
      repbyTime(FileInput, FileOutput, Srep=18.5, me) 
    }
}
\author{
Xiaosu Tong
}

