% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetime.control.R
\name{spacetime.control}
\alias{spacetime.control}
\title{Set parameter for drSpaceTime fitting}
\usage{
spacetime.control(inner, outer, surface = c("interpolate", "direct"),
  lib.loc = NULL, iterations = 1L)
}
\arguments{
\item{inner}{The iteration time for inner loop of stlplus for time dimension fitting}

\item{surface}{should the fitted surface be computed exactly or via interpolation from a kd tree?}

\item{lib.loc}{the library searching path, sepcifically the path including Spaloess and 
stlplus package on the front end server. If all packages have been pushed
the the tar.gz file on hdfs, then this argument is set to be NULL.}

\item{iterations}{the number of iterations used for the space-time back-fitting.}

\item{output}{The iteration time for outer loop of stlplus for time dimension fitting}
}
\value{
A list with space-time fitting parameters.
}
\description{
Set control parameters for drSpaceTime fits
}
\examples{
    FileInput <- "/ln/tongx/Spatial/a1950/bystation"
    FileOutput <- "/ln/tongx/Spatial/a1950/bymonth"
    \dontrun{
      swaptoTime(FileInput, FileOutput, elevFlag=TRUE)
    }
}
\author{
Xiaosu Tong
}

