% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetime.control.R
\name{spacetime.control}
\alias{spacetime.control}
\title{Set smoothing parameter for drSpaceTime fitting}
\usage{
spacetime.control(vari, time, seaname, n, n.p, s.window, s.degree = 1,
  s.jump = ceiling(s.window/10), t.window = NULL, t.degree = 1,
  t.jump = ceiling(t.window/10), t.blend = 0, l.window = NULL,
  l.degree = t.degree, l.jump = ceiling(l.window/10), l.blend = 0,
  critfreq = 0.05, s.blend = 0, sub.labels = NULL, sub.start = 1,
  inner = 4, outer = 1, surface = c("interpolate", "direct"),
  iterations = 1, infill = TRUE)
}
\arguments{
\item{vari}{variable name in string of the response variable}

\item{time}{variable name in string of time index of the whole time series}

\item{seaname}{variable name in string of the seasonal variable}

\item{n}{the number of total observations}

\item{n.p}{the number of observation in each subseries}

\item{s.window}{either the character string \code{"periodic"} or the span (in lags) of the loess window for seasonal extraction, which should be odd. This has no default.}

\item{s.degree}{degree of locally-fitted polynomial in seasonal extraction. Should be 0, 1, or 2.}

\item{s.jump, t.jump, l.jump}{integers at least one to increase speed of the respective smoother. Linear interpolation happens between every \code{*.jump}th value.}

\item{s.blend, t.blend, l.blend}{vectors of proportion of blending to degree 0 polynomials at the endpoints of the series.}

\item{sub.labels}{optional vector of length n.p that contains the labels of the subseries in their natural order (such as month name, day of week, etc.), used for strip labels when plotting. All entries must be unique.}

\item{sub.start}{which element of sub.labels does the series begin with. See details.}

\item{inner}{The iteration time for inner loop of stlplus for time dimension fitting}

\item{outer}{The iteration time for outer loop of stlplus for time dimension fitting}

\item{surface}{should the fitted surface be computed exactly or via interpolation from a kd tree?}

\item{iterations}{the number of iterations used for the space-time back-fitting.}

\item{zero.weight}{value to use as zero for zero weighting}

\item{libLoc}{the library searching path, sepcifically the path including Spaloess and 
stlplus package on the front end server. If all packages have been pushed
the the tar.gz file on hdfs, then this argument is set to be NULL.}
}
\value{
A list with space-time fitting parameters.
}
\description{
Set control parameters for drSpaceTime fits
}
\examples{
    spacetime.control()
}
\author{
Xiaosu Tong
}

