% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialfit.R
\name{spatialfit}
\alias{spatialfit}
\title{Conduct the stlplus fitting at each location in parallel}
\usage{
spatialfit(input, output, info, model_control = spacetime.control(),
  cluster_control = mapreduce.control())
}
\arguments{
\item{input}{The path of input sequence file on HDFS. It should be by location division.}

\item{output}{The path of output sequence file on HDFS. It is by location division but with seasonal and trend components}

\item{model_control}{The list contains all smoothing parameters}

\item{cluster_control}{A list contains all mapreduce tuning parameters.}
}
\description{
call stlplus function on time series at each location in parallel.
Every station uses the same smoothing parameter
}
\examples{
    FileInput <- "/wsc/tongx/spatem/tmax/sim/byyr256"
    FileOutput <- "/wsc/tongx/spatem/tmax/sim/byyrfit256"
    me <- mapreduce.control(libLoc=lib.loc, BLK = 256)
    you <- spacetime.control(vari="resp", time="date", seaname="month", n=4448736, n.p=12, s.window=13, t.window = 241, degree=2, span=0.015, Edeg=2)
    \dontrun{
      spatialfit(FileInput, FileOutput, info="/wsc/tongx/spatem/stationinfo/UStinfo.RData", you, me)
    }
}
\author{
Xiaosu Tong
}

