% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstl_local.R
\name{sstl_local}
\alias{sstl_local}
\title{Readin Raw text data files and save it as by time division on HDFS.}
\usage{
sstl_local(data, model_control = spacetime.control())
}
\arguments{
\item{data}{The input data.frame which contains observation, lon, lat, elev, and year, month}

\item{model_control}{Should be a list object generated from \code{spacetime.control} function.
The list including all necessary smoothing parameters of nonparametric fitting.}
}
\description{
Input raw text data file is divided into by time subsets and saved on HDFS
}
\examples{
    head(tmax_all)
    mcontrol <- spacetime.control(
      vari="tmax", n=576, n.p=12, stat_n=7738,
      s.window=13, t.window = 241, degree=2, span=0.015, Edeg=2
    )
    sstl_local(tmax_all, model_control=mcontrol)
}
\author{
Xiaosu Tong
}

