% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stlfit.R
\name{stlfit}
\alias{stlfit}
\title{Apply the stlplus fitting at each location in parallel}
\usage{
stlfit(input, output, model_control = spacetime.control(),
  cluster_control = mapreduce.control())
}
\arguments{
\item{input}{The path of input sequence file on HDFS. It should be by location division.}

\item{output}{The path of output sequence file on HDFS. It is by location division but with seasonal and trend components}

\item{model_control}{The list contains all smoothing parameters}

\item{cluster_control}{A list contains all mapreduce tuning parameters.}
}
\description{
Calling \code{stlplus} function from Ryan Hafen's \code{stlplus}
package on time series at each location in parallel.
Every station uses the same smoothing parameter
}
\examples{
\dontrun{
    FileInput <- "/tmp/bystat"
    FileOutput <- "/tmp/bystatfit"
    ccontrol <- mapreduce.control(libLoc=NULL, reduceTask=0)
    mcontrol <- spacetime.control(
      vari = "resp", time = "date", n = 576, stat_n=7738, n.p = 12, s.window = "periodic",
      t.window = 241, degree = 2, span = 0.015, Edeg = 2
    )
    stlfit(FileInput, FileOutput, model_control=mcontrol, cluster_control=ccontrol)
}
}
\author{
Xiaosu Tong
}
\references{
R. B. Cleveland, W. S. Cleveland, J. E. McRae, and I. Terpenning (1990) STL: A Seasonal-Trend Decomposition Procedure Based on Loess. \emph{Journal of Official Statistics}, \bold{6}, 3--73.

Ryan Hafen (2010): stlplus: Local regression models: Advancements, applications, and new methods. \emph{Purdue University}
}
\seealso{
\code{\link{spacetime.control}}, \code{\link{mapreduce.control}}
}

