% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swaptoTime.R
\name{swaptoTime}
\alias{swaptoTime}
\title{Swap to division by time}
\usage{
swaptoTime(input, output, control = mapreduce.control())
}
\arguments{
\item{input}{The path of input sequence file on HDFS. It should be by location division.}

\item{output}{The path of output sequence file on HDFS. It is by time division.}

\item{control}{all parameters that are needed for space-time fitting}

\item{reduceTask}{The number of the reduce tasks.}

\item{elevFlag}{Logical argument, if TRUE, then the elevation attribute from input value data.frame is log2 transformation
If FALSE, a log2 transformation is added to the elevation attribute.}
}
\description{
Switch input key-value pairs which is division by location
to the key-value pairs which is division by time.
}
\details{
swaptoTime is used for switch division by location to division by time.
}
\examples{
    FileInput <- "/wsc/tongx/Spatial/tmp/tmax/simulate/bystation"
    FileOutput <- "/wsc/tongx/Spatial/tmp/tmax/simulate/bymonth"
    me <- mapreduce.control(libLoc=lib.loc, io_sort=2047)
    \dontrun{
      swaptoTime(FileInput, FileOutput, me)
    }
}
\author{
Xiaosu Tong
}

