% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swaptoTime.R
\name{swaptoTime}
\alias{swaptoTime}
\title{Swap to division by time}
\usage{
swaptoTime(input, output)
}
\arguments{
\item{input}{The path of input sequence file on HDFS. It should be by location division.}

\item{output}{The path of output sequence file on HDFS. It is by time division.}

\item{reduceTask}{The number of the reduce tasks.}

\item{elevFlag}{Logical argument, if TRUE, then the elevation attribute from input value data.frame is log2 transformation
If FALSE, a log2 transformation is added to the elevation attribute.}

\item{control}{all parameters that are needed for space-time fitting}
}
\description{
Switch input key-value pairs which is division by location
to the key-value pairs which is division by time.
}
\details{
swaptoTime is used for switch division by location to division by time.
}
\examples{
    FileInput <- "/wsc/tongx/spatem/tmax/sim/bystatfit512"
    FileOutput <- "/wsc/tongx/spatem/tmax/sim/byyear"
    me <- mapreduce.control(libLoc="/home/tongx/R_LIBS", io_sort=512, BLK=256, reduce_input_buffer_percent=0.7, reduce_merge_inmem=0, task_io_sort_factor=100, spill_percent=1)
    \dontrun{
      swaptoTimeMap(FileInput, FileOutput, me)
      swaptoTimeRed(FileOutput, "/wsc/tongx/spatem/tmax/sim/byyr256test", me)
    }
}
\author{
Xiaosu Tong
}

