% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drsstl-package.R
\docType{data}
\name{tmax_all}
\alias{tmax_all}
\title{US Maximum Temperature}
\source{
Full data on NCAR: \url{http://www.image.ucar.edu/Data/US.monthly.met/FullData.shtml#temp}
}
\description{
A data.frame contains maximum temperature data from NCAR weside
}
\keyword{data}

